:rocket: new commit 20250424 /  Order 변경 - 예외처리 추가

+++ src/main/java/com/grepp/spring/app/model/order/OrderService.java
@@ -1,6 +1,7 @@
 package com.grepp.spring.app.model.order;
 
 import com.grepp.spring.app.model.order.dto.Order;
+import com.grepp.spring.infra.error.exceptions.OrderNotFoundException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,11 +15,19 @@ public class OrderService {
     private final OrderRepository orderRepository;
 
     public List<Order> getOrdersById(String id) {
-        return orderRepository.findById(id);
+        List<Order> orders = orderRepository.findById(id);
+        if (orders == null || orders.isEmpty()) {
+            throw new OrderNotFoundException("해당 ID로 주문을 찾을 수 없습니다: " + id);
+        }
+        return orders;
     }
 
     public List<Order> getOrdersByEmail(String email) {
-        return orderRepository.findByEmail(email);
+        List<Order> orders = orderRepository.findById(email);
+        if (orders == null || orders.isEmpty()) {
+            throw new OrderNotFoundException("해당 ID로 주문을 찾을 수 없습니다: " + email);
+        }
+        return orders;
     }
 
     public List<Order> getAllOrders() {
@@ -26,6 +35,10 @@ public class OrderService {
     }
 
     public void deleteOrder(String email) {
+        List<Order> orders = orderRepository.findById(email);
+        if (orders == null || orders.isEmpty()) {
+            throw new OrderNotFoundException("삭제할 주문이 없습니다.: " + email);
+        }
         orderRepository.deleteByEmail(email);
     }
